rustc_span/symbol.rs Code Quality Impr.
The struct ‘InternerInner’ should not be holding static references to strings, as this will mislead future contributors. This will obstruct future progress, albeit working out right now.

The idiomatic implementation to represent a low-level struct ‘owning’ some sort of untyped data is to use raw-pointer-fields (see example for ‘PhantomData’ in std) and only convert them into references inside methods of the struct. This better reflects that references to the data can only be safely obtained via the struct’s carefully crafted methods. The current implementation suggests otherwise; namely that the '’InternerInner’'s references can be freely assumed to be soundly used as ‘'static’, if one overlooks the provided comment. This increases the chance for accidental unsound implementations and hinders effectivity the more the code’s requirements become complex.

For example, disregarding the use-case-plausibility of the following scenario, an implementor could create a method on ‘InternerInner’ to replace its ‘DroplessArena’ from whatever reason. As of now, he would do this without second thoughts. My suggestions applied, the struct’s raw pointers would highlight and connote that certain safety requirements are.