rustdoc: use browser-native tooltips for code and notable hover
This is the simplest way to adopt the timing guidelines described
in <https://www.nngroup.com/articles/timing-exposing-content/>.
It replaces the mouseover interactions used for "tooltip" popovers
with HTML `title` attributes, which already do the right thing.

The JavaScript-based popover code remains for click interaction, so
that these elements remain accessible on touch screens, and because
it's not possible to put links in native tooltips.