Add a feature to remove the static initializer on Linux/GNU
The `explicit-init-args-with-glibc` feature will disable
the static initializer that is used on Linux with glibc
to initialize `std::env::args`. The result is that the
args will be empty in a cdylib on this platform.

The feature is off by default, so there is no change to
any Rust users unless they build stdlib themselves and
turn the feature on explicitly.