Remove `MonoItems`.
PR #97168 was created to address this comment in `record_accesses`:

// We collect this into a `SmallVec` to avoid calling `is_inlining_candidate` in the lock.
// FIXME: Call `is_inlining_candidate` when pushing to `neighbors` in `collect_items_rec`
// instead to avoid creating this `SmallVec`.

It removed the need for the `SmallVec` (a good thing) by introducing
`MonoItems`, which introduced some additional complexity (a bad thing).

However, the reason the `SmallVec` was being used in the first place --
"to avoid calling `is_inlining_candidate` in the lock" -- isn't
necessary, as far as I can tell. Perhaps it was at one point.

So this commit just moves what is left of the call to
`is_inlining_candidate` -- a call the innocuous `instantiation_mode` --
inside the lock. This avoids the complexity of `MonoItems` sometimes
computing inlining data and sometimes not.