Disable the `BorrowedFd` layout optimization on Haiku
On Haiku (and apparently only Haiku), the platform `AT_FDCWD` has the value
`-1`. However, `BorrowedFd` currently has layout optimizations that assume the
value is not `-1`, and it's useful to be able to store `AT_FDCWD` in a
`BorrowedFd`. For example, rustix has
[a function which returns a `BorrowedFd` containing `AT_FDCWD`].

To support this on Haiku, disable the layout optimizations for `BorrowedFd` on
Haiku and allow it to store the value `-1`.

This is technically a breaking change for Haiku users, as they could
theoretically be depending on `Option<BorrowedFd>` having the same layout as
`RawFd`, however this is not a common idiom, and Haiku support is [Tier 3]
in Rust.

[a function which returns a `BorrowedFd` containing `AT_FDCWD`]: https://docs.rs/rustix/latest/rustix/fs/fn.cwd.html
[Tier 3]: https://doc.rust-lang.org/nightly/rustc/platform-support.html#tier-3