Avoid `memcpy` in codegen for more types, notably `Vec`
PR 111999 set up the framework to be able to do this; this PR expands it to more types than just arrays.

Most interestingly, this allows it to work with `Vec<T>` and `String`, so swapping those no longer ends up going through stack like it does today (<https://rust.godbolt.org/z/cKG7o8aaW>).

And since this is done in codegen, it's not special for `swap`, and thus will hopefully allow types like this to better optimize in lots of places, with easier SRoA.