Create `implicit_transmute_types` lint.
Type inference is useful in most cases, but it can cause issues around usage of transmute.
Even though a specific usage of transmute may have been confirmed to be sound, someone might
unintentionally change a piece of surrounding code that causes the transmuted types to
change and possibly become unsound. Despite that, transmute() will happily continue to infer these
changed types.

This commit adds a lint `implicit_transmute_types` that catches calls to transmute without
an explicit turbofish. The idea is that the initial author of code that uses a transmute will
verify the soundness of converting between the two types, and enfore those types on the transmute
call. If the types end up changing later, typecheck will fail and force the transmute call to be
updated and remind the user to reassess the soundness of transmuting between the new types.

The lint is "allow" by default because it would trigger in lots of existing rust code, including
many cases in the core libraries. It may be beneficial to make it deny-by-default in the future
though.